requirements.txt
absl-py==2.2.0
accelerate==1.4.0
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
akshare==1.17.4
alembic==1.15.2
altair==5.5.0
annotated-types==0.7.0
anthropic==0.18.1
anyio==4.9.0
anytree==2.12.1
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
arxiv==2.1.3
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==4.0.3
asyncache==0.3.1
asyncer==0.0.8
attrs==25.1.0
babel==2.17.0
banal==1.0.6
beautifulsoup4==4.13.3
bert-score==0.3.13
bioc==2.1
biopython==1.85
bleach==6.2.0
blinker==1.9.0
boto3==1.37.28
botocore==1.37.28
Brotli @ file:///C:/b/abs_c415aux9ra/croot/brotli-split_1736182803933/work
cachetools==5.5.2
certifi @ file:///C:/b/abs_3beajm7umk/croot/certifi_1745939228545/work/certifi
cffi==1.17.1
chardet==5.2.0
charset-normalizer @ file:///croot/charset-normalizer_1721748349566/work
click==8.1.8
cloudpickle==3.1.1
cloudscraper==1.2.71
colorama==0.4.6
comm==0.2.2
conda-pack @ file:///C:/b/abs_0cu91sbau5/croot/conda-pack_1710258097566/work
conllu==4.5.3
contourpy==1.3.1
cycler==0.12.1
dashscope==1.22.2
dataclasses-json==0.6.7
dataset==1.6.2
datasets==3.3.2
ddg==0.2.2
debugpy==1.8.12
decorator==5.2.1
deep-translator==1.11.4
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
docopt==0.6.2
dotenv==0.9.9
dropbox==12.0.2
duckduckgo_search==5.3.1b1
editdistance==0.8.1
einops==0.8.1
elsapy==0.5.1
et_xmlfile==2.0.0
exceptiongroup==1.2.2
executing==2.2.0
fastapi==0.115.10
fastjsonschema==2.21.1
feedparser==6.0.11
ffmpy==0.5.0
filelock @ file:///C:/b/abs_f2gie28u58/croot/filelock_1700591233643/work
fire==0.7.0
flair==0.15.1
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
ftfy==6.3.1
func_timeout==4.3.5
gast==0.6.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///C:/b/abs_d8ki0o0h97/croot/gmpy2_1738085498525/work
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc1
google-api-python-client==2.170.0
google-auth==2.40.2
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
google-pasta==0.2.0
google_search_results==2.4.2
googleapis-common-protos==1.70.0
gradio==5.7.1
gradio_client==1.5.0
greenlet==3.1.1
griffe==0.49.0
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
h2==4.2.0
h5py==3.13.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.4
hyperframe==6.1.0
idna==3.10
interegular==0.3.3
intervaltree==3.1.0
ipykernel==6.29.5
ipython==8.33.0
ipywidgets==8.1.5
isoduration==20.11.0
janus==2.0.0
jedi==0.19.2
jieba==0.42.1
Jinja2 @ file:///C:/b/abs_10zs3i4lof/croot/jinja2_1737760123028/work
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
json5==0.10.0
jsonlines==4.0.0
jsonpatch==1.33
jsonpath==0.82.2
jsonpickle==4.0.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.12.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
kiwisolver==1.4.8
lagent==0.5.0rc2
langchain==0.3.20
langchain-community==0.3.19
langchain-core==0.3.43
langchain-text-splitters==0.3.6
langdetect==1.0.9
langsmith==0.3.13
language_tool_python==2.9.2
lark==1.2.2
lemminflect==0.2.3
libclang==18.1.1
llvmlite==0.44.0
lmdeploy==0.7.1
lru-dict==1.3.0
lxml==5.3.2
Mako==1.3.10
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.1
matplotlib-inline==0.1.7
mcp==1.6.0
mdurl==0.1.2
mini-racer==0.12.4
mistune==3.1.2
mkl-service==2.4.0
mkl_fft @ file:///C:/Users/dev-admin/mkl/mkl_fft_1730823082242/work
mkl_random @ file:///C:/Users/dev-admin/mkl/mkl_random_1730822522280/work
ml_dtypes==0.5.1
mmengine-lite==0.10.6
more-itertools==10.6.0
mpld3==0.5.10
mpmath @ file:///C:/b/abs_7833jrbiox/croot/mpmath_1690848321154/work
multidict==6.1.0
multiprocess==0.70.16
mypy-extensions==1.0.0
namex==0.0.8
narwhals==1.28.0
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx @ file:///C:/b/abs_b054htfn9t/croot/networkx_1737043671910/work
nltk==3.9.1
notebook==7.3.2
notebook_shim==0.2.4
num2words==0.5.14
numba==0.61.0
numpy==1.26.4
ollama==0.4.7
openai==1.66.3
OpenHowNet==2.0
openpyxl==3.1.5
opt_einsum==3.4.0
optree==0.14.1
orjson==3.10.15
outcome==1.3.0.post0
outlines==0.0.46
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
peft==0.14.0
pillow @ file:///C:/b/abs_b50vowcrzo/croot/pillow_1738010273782/work
pinyin==0.4.0
platformdirs==4.3.6
playwright==1.51.0
ply==3.11
pptree==3.1
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
pure_eval==0.2.3
pyairports==2.1.1
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycountry==24.6.1
pycparser==2.22
pydantic==2.6.4
pydantic-settings==2.8.1
pydantic_core==2.16.3
pydeck==0.9.1
pydub==0.25.1
pyee==12.1.1
Pygments==2.19.1
pyparsing==3.2.1
PySocks @ file:///C:/ci_310/pysocks_1642089375450/work
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
python-multipart==0.0.12
pytorch_revgrad==0.2.0
pytz==2025.1
pyvis==0.3.2
pywin32==308
pywinpty==2.0.15
PyYAML @ file:///C:/b/abs_14xkfs39bx/croot/pyyaml_1728657968772/work
pyzmq==26.2.1
qtconsole==5.6.1
QtPy==2.4.3
referencing==0.36.2
regex==2024.11.6
requests @ file:///C:/b/abs_c3508vg8ez/croot/requests_1731000584867/work
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9.1
ruff==0.9.9
s3transfer==0.11.4
safehttpx==0.1.6
safetensors==0.5.3
schemdraw==0.19
scikit-learn==1.6.1
scipy==1.15.2
segtok==1.5.11
selenium==4.31.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==4.1.0
sentencepiece==0.2.0
sgmllib3k==1.0.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==1.4.54
sqlitedict==2.1.0
sse-starlette==2.2.1
stack-data==0.6.3
starlette==0.46.0
stone==3.3.1
streamlit==1.42.2
sympy==1.13.1
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
terminado==0.18.1
terminaltables==3.1.10
textattack==0.3.10
textblob==0.19.0
threadpoolctl==3.6.0
tiktoken==0.9.0
timeout-decorator==0.5.0
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
touch==2020.12.3
tqdm==4.67.1
traitlets==5.14.3
transformer-smaller-training-vocab==0.4.0
transformers==4.52.2
trio==0.29.0
trio-websocket==0.12.2
typer==0.15.2
types-python-dateutil==2.9.0.20241206
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2025.1
undetected-chromedriver==3.5.5
uri-template==1.3.0
uritemplate==4.1.1
urllib3 @ file:///C:/b/abs_7bst06lizn/croot/urllib3_1737133657081/work
uvicorn==0.34.0
watchdog==6.0.0
wcwidth==0.2.13
webcolors==24.11.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==12.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
Wikipedia-API==0.8.1
win-inet-pton @ file:///C:/ci_310/win_inet_pton_1642658466512/work
word2number==1.1
wrapt==1.17.2
wsproto==1.2.0
xlrd==2.0.2
xxhash==3.5.0
yapf==0.43.0
yarl==1.18.3
zstandard==0.23.0
